+++ src/components/Home/index.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import Navbar from "@/components/Navbar";
 import Hero from "@/components/Hero";
-import Search from "@/components/Search";
 import { mockListings } from "@/lib/mockData";
 import Reels from "@/components/Reels";
 import FeaturedProperties from "@/components/FeaturedProperties";
@@ -9,14 +8,17 @@ import CollapsibleMap from "@/components/Map/collapsibleMap";
 import LazySection from "@/components/LazySection";
 import Footer from "@/components/Footer";
 import Blogs from "@/components/Blogs";
-
+import { HomeListingsSection } from "@/components/HomeListingsSection";
+import { AuthGuard } from "../Auth/authGuard";
 export default function Home() {
   return (
     <>
       <Navbar />
       <Hero />
       <LazySection>
-        <Search />
+        <AuthGuard renderPrivate={false}>
+          {(user, loginUI) => <HomeListingsSection user={user} />}
+        </AuthGuard>
       </LazySection>
 
       <LazySection>