+++ CHANGELOG.md
@@ -25,3 +25,8 @@
 - Implemented property detail page
 - Implemented RETS Script
 - Implemented listing detail page
+- Added Filter by market status on Map
+- Implemented Featured Property Component and Search Component on Home Page
+- Implemented All Listings Page with Filters and Pagination ; added search properies only option
+- Added Recent Solds and Latest Listings Component under Property Detail Page
+- Modified Featured Property Management under Admin